chore(demo): update new demo structure with changes from development
