+++ demo/src/app/demo/components/demo.component.html
@@ -1,10 +1,7 @@
-<div class="container" highlight-js-content=".highlight">
+<div class="container mb-5" highlight-js-content=".highlight">
   <basic></basic>
-  <div class="mt-3"></div>
   <filter></filter>
-  <div class="mt-3"></div>
   <customise></customise>
-  <div class="mt-3"></div>
   <events></events>
-  <div class="mt-5"></div>
+  <styles></styles>
 </div>
