+++ demo/src/app/demo/components/style/style.component.html
@@ -6,22 +6,26 @@
 
 <pre><code class="language-markup"><![CDATA[<image-upload url="https://httpbin.org/status/200" class="customClass"></image-upload>]]></code></pre>
 
-<pre><code class="language-css"><![CDATA[.customClass{    
+<pre><code class="language-css"><![CDATA[.customClass {
     background-color: #dd3;
     border-radius: 5px;
     margin:5px;
-    width: 500px;        
+    width: 500px;
 }
-.customClass .img-ul-upload{
+
+.customClass .img-ul-upload {
     background-color: #000 !important;
 }
-.customClass .img-ul-clear{
+
+.customClass .img-ul-clear {
     background-color: #B819BB !important;
 }
+
 .customClass .img-ul-drag-box-msg {
     color: purple !important;
 }
-.customClass .img-ul-container{
+
+.customClass .img-ul-container {
     background-color: #FF6CAD !important;
 }
 ]]></code></pre>
@@ -38,7 +42,7 @@
       "border-radius": "25px",
       "color": "#000"
     },
-    clearButton: {      
+    clearButton: {
       "background-color": "#FFF",
       "border-radius": "25px",
       "color": "#000",
@@ -53,7 +57,7 @@
       "padding-top": "5px",
       "width": "500px"
     },
-    previewPanel: {      
+    previewPanel: {
       "background-color": "#894489",
       "border-radius": "0 0 25px 25px",
     }
