--- demo/src/app/style/style.component.css
@@ -1,18 +1,0 @@
-.customClass{    
-    background-color: #dd3;
-    border-radius: 5px;
-    margin:5px;
-    width: 500px;        
-}
-.customClass .img-ul-upload{
-    background-color: #000 !important;
-}
-.customClass .img-ul-clear{
-    background-color: #B819BB !important;
-}
-.customClass .img-ul-drag-box-msg {
-    color: purple !important;
-}
-.customClass .img-ul-container{
-    background-color: #FF6CAD !important;
-}
