+++ demo/src/assets/readme.md
@@ -1,6 +1,6 @@
 [![Build Status](https://travis-ci.org/aberezkin/ng2-image-upload.svg?branch=master)](https://travis-ci.org/aberezkin/ng2-image-upload)
 
-This angular 2 library provides a light-weight component that handles file-drop, image previewing and image uploading.
+This angular library provides a light-weight component that handles file-drop, image previewing and image uploading.
 
 ### Install
 
@@ -10,6 +10,8 @@ This angular 2 library provides a light-weight component that handles file-drop,
 
 In your `app.module.ts` import it using `@NgModule` decorator.
 
+`import { ImageUploadModule } from "angular2-image-upload";`
+  
 ````typescript
 @NgModule({
     imports: [
@@ -19,6 +21,7 @@ In your `app.module.ts` import it using `@NgModule` decorator.
     ]
 })
 ````
+
 Now you have `image-upload` declaration and you can use it in your html code.
 
 `<image-upload></image-upload>`
@@ -46,10 +49,10 @@ If you need to send some headers with your request (for example `Authorization` 
 you can use `[headers]` directive like this.
 
 ````html
-<image-upload [url]="'my-url.com'"
-  [headers]="[
-    {header: 'Authorization', value: 'MyToken'}
-  ]"></image-upload>
+<image-upload 
+  [url]="'my-url.com'"
+  [headers]="{Authorization: 'MyToken'}">
+</image-upload>
 ````
 
 **Note** that headers are sent only if you provide a url.
@@ -62,13 +65,79 @@ you can use `[headers]` directive like this.
 
 `[fileTooLargeMessage]="'Image too large!'"` - message that is shown if the user selects/drops an image that exceeds `maxFileSize`. Default is "**An image was too large and was not uploaded. The maximum file size is x KiB.**".
 
-#### Callbacks
+`[clearButtonCaption]="'Clear'"` - Text shown on the "Clear" button. Default is "**Clear**".
 
-`(onFileUploadFinish)="imageUploaded($event)"`. If `[url]` is specified this event is fired when component gets a response from the server, also in this case event has field `serverResponse` which contains the status code and response from the server `{status, response}`. If `[url]` is not specified it's fired immediately after an image(s) dropped into file-drop zone of choosed in file browser. So what you can do, is not specify `[url]` to handle upload yourself, for exapmple send the image into firebase storage. To get file use `event.file`.
+#### CSS Class
 
-`(onRemove)="imageRemoved($event)"` - this event is fired when remove button was clicked and the image preview was removed. *Note that this library doesn't handle deletion from server so you should do it yourself*. Event passed as the argument is the exact same object that was passed to the `(imageUploaded)` callback when image was added so you can access `serverResponse` to get a key to delete your image from server.
+`[class]="'customClass'"` - Set custom class for this component. It is set on parent level element `<image-upload>` which can be customized as follow :
 
-`(isPending)="disableSendButton($event)"` - this event is fired when pending state was changed. Event is just a boolean that represents the pending state. Pending state is `true` when and only when component avaits a response from the server, and `false` othervise. You can use it, for example, to disable send button in your form until all images are uploaded.
+````css
+.customClass {
+    background-color: #dd3;
+    border-radius: 5px;
+    margin:5px;
+    width: 500px;
+}
+
+.customClass .img-ul-upload {
+    background-color: #000 !important;
+}
+
+.customClass .img-ul-clear {
+    background-color: #B819BB !important;
+}
+
+.customClass .img-ul-drag-box-msg {
+    color: purple !important;
+}
+
+.customClass .img-ul-container {
+    background-color: #FF6CAD !important;
+}
+````
+
+
+**The class must be accessible to the ImageUploadComponent**, so either pop it in your global stylesheet, or if you need to put it in a scoped stylesheet prefix with `/deep/`:
+
+`/deep/ .customClass { ... }`
+    
+**Note:**  `.img-ul-*` classes which are overridden with new styles. 
+
+#### Custom Style
+
+
+`[style]="'customStyle'"` - Set custom style properties for this component. `customStyle` is typescript object defined in your component.
+
+````typescript
+customStyle = {
+  selectButton: {
+    "color": "white",
+    "background-color": "purple",
+  },
+  clearButton: {
+    "color": "white",
+    "background-color": "yellow",
+  },
+  layout: {
+    "background-color": "black",
+    "color": "red",
+    "font-size": "15px",
+  },
+  previewPanel: {
+    "background-color": "red",
+  }
+};
+````
+      
+**Note:** `selectButton`, `clearButton`, `layout` and `previewPanel` are optional properties.
+
+#### Events
+
+`(uploadFinished)="onUploadFinished($event)"`. If `[url]` is specified this event is fired when component gets a response from the server, also in this case event has field `serverResponse` which contains the status code and response from the server `{status, response}`. If `[url]` is not specified it's fired immediately after an image(s) dropped into file-drop zone of choosed in file browser. So what you can do, is not specify `[url]` to handle upload yourself, for exapmple send the image into firebase storage. To get file use `event.file`.
+
+`(removed)="onRemoved($event)"` - this event is fired when remove or clear button was clicked and the image preview was removed. *Note that this library doesn't handle deletion from server so you should do it yourself*. Event passed as the argument is the exact same object that was passed to the `(imageUploaded)` callback when image was added so you can access `serverResponse` to get a key to delete your image from server.
+
+`(uploadStateChanged)="onUploadStateChanged($event)"` - this event is fired when image upload state was changed. Event is just a boolean that represents the uploading state. Image upload state is `true` when and only when component awaits the response from the server, and `false` otherwise. You can use it, for example, to disable send button in your form until all images are uploaded.
 
 In the final state it should look something like this:
 
@@ -76,17 +145,17 @@ In the final state it should look something like this:
 <image-upload
   [max]="100"
   [url]="'example.com/images/upload'"
-  [headers]="[
-    {header: 'Authorization', value: 'MyToken'}
-  ]"
+  [headers]="{Authorization: 'MyToken'}"
   [buttonCaption]="'Select Images!'"
   [dropBoxMessage]="'Drop your images here!'"
   [extensions]="['jpg','png','gif']"
-  (onFileUploadFinish)="imageUploaded($event)"
-  (onRemove)="imageRemoved($event)"
-  (isPending)="disableSendButton($event)"
-></image-upload>
+  [class]="'customClass'"      
+  (removed)="onRemoved($event)"
+  (uploadFinished)="onUploadFinished($event)"
+  (uploadStateChanged)="onUploadStateChanged($event)">
+</image-upload>
 ````
+
 # Contributors
 
 @aberezkin
