+++ demo/src/styles.css
@@ -21,3 +21,26 @@ h1.root-loading {
 .token.operator {
   background: none !important;
 }
+
+.customClass {
+  background-color: #dd3;
+  border-radius: 5px;
+  margin: 5px;
+  width: 500px;
+}
+
+.customClass .img-ul-upload {
+  background-color: #000 !important;
+}
+
+.customClass .img-ul-clear {
+  background-color: #B819BB !important;
+}
+
+.customClass .img-ul-drag-box-msg {
+  color: purple !important;
+}
+
+.customClass .img-ul-container {
+  background-color: #FF6CAD !important;
+}
